requirements.txt
faiss-cpu
sentence-transformers
fastapi
uvicorn
pandas
numpy
scikit-learn
PyPDF2
tqdm



